move test to parent so as to test oracle query store
