+++ ledger-service/http-json/src/itlib/scala/http/AbstractHttpServiceIntegrationTest.scala
@@ -1822,4 +1822,110 @@ abstract class AbstractHttpServiceIntegrationTest
       _ <- queryN(0)
     } yield succeed
   }
+
+  "Should ignore conflicts on contract key hash constraint violation" in withHttpServiceAndClient {
+    (uri, encoder, _, _, _) =>
+      import scalaz.std.vector._
+      import scalaz.syntax.tag._
+      import scalaz.syntax.traverse._
+      import scalaz.std.scalaFuture._
+      import shapeless.record.{Record => ShRecord}
+      import scalaz.std.option.some
+      import com.daml.ledger.api.refinements.{ApiTypes => lar}
+
+      val partyIds = Vector("Alice", "Bob").map(getUniqueParty)
+      val packageId: Ref.PackageId = MetadataReader
+        .templateByName(metadataUser)(Ref.QualifiedName.assertFromString("User:User"))
+        .collectFirst { case (pkgid, _) => pkgid }
+        .getOrElse(fail(s"Cannot retrieve packageId"))
+
+      def userCreateCommand(
+          username: domain.Party,
+          following: Seq[domain.Party] = Seq.empty,
+      ): domain.CreateCommand[v.Record, domain.TemplateId.OptionalPkg] = {
+        val templateId = domain.TemplateId(None, "User", "User")
+        val followingList = following.map(party => v.Value(v.Value.Sum.Party(party.unwrap)))
+        val arg = v.Record(
+          fields = List(
+            v.RecordField("username", Some(v.Value(v.Value.Sum.Party(username.unwrap)))),
+            v.RecordField(
+              "following",
+              some(v.Value(v.Value.Sum.List(v.List.of(followingList)))),
+            ),
+          )
+        )
+
+        domain.CreateCommand(templateId, arg, None)
+      }
+      def userExerciseFollowCommand(
+          contractId: lar.ContractId,
+          toFollow: domain.Party,
+      ): domain.ExerciseCommand[v.Value, domain.EnrichedContractId] = {
+        val templateId = domain.TemplateId(None, "User", "User")
+        val reference = domain.EnrichedContractId(Some(templateId), contractId)
+        val arg = recordFromFields(ShRecord(userToFollow = v.Value.Sum.Party(toFollow.unwrap)))
+        val choice = lar.Choice("Follow")
+
+        domain.ExerciseCommand(reference, choice, boxedRecord(arg), None)
+      }
+
+      def followUser(contractId: lar.ContractId, actAs: domain.Party, toFollow: domain.Party) = {
+        val exercise: domain.ExerciseCommand[v.Value, domain.EnrichedContractId] =
+          userExerciseFollowCommand(contractId, toFollow)
+        val exerciseJson: JsValue = encodeExercise(encoder)(exercise)
+
+        postJsonRequest(
+          uri.withPath(Uri.Path("/v1/exercise")),
+          exerciseJson,
+          headers = headersWithPartyAuth(actAs = List(actAs.unwrap)),
+        )
+          .map { case (exerciseStatus, exerciseOutput) =>
+            exerciseStatus shouldBe StatusCodes.OK
+            assertStatus(exerciseOutput, StatusCodes.OK)
+            ()
+          }
+
+      }
+
+      def queryUsers(fromPerspectiveOfParty: domain.Party) = {
+        val query = jsObject(s"""{
+             "templateIds": ["$packageId:User:User"],
+             "query": {}
+          }""")
+
+        postJsonRequest(
+          uri.withPath(Uri.Path("/v1/query")),
+          query,
+          headers = headersWithPartyAuth(actAs = List(fromPerspectiveOfParty.unwrap)),
+        ).map { case (searchStatus, searchOutput) =>
+          searchStatus shouldBe StatusCodes.OK
+          assertStatus(searchOutput, StatusCodes.OK)
+        }
+      }
+      val commands = partyIds.map { p =>
+        (p, userCreateCommand(p))
+      }
+
+      for {
+        users <- commands.traverse { case (party, command) =>
+          val fut = postCreateCommand(
+            command,
+            encoder,
+            uri,
+            headers = headersWithPartyAuth(actAs = List(party.unwrap)),
+          ).map { case (status, output) =>
+            status shouldBe StatusCodes.OK
+            assertStatus(output, StatusCodes.OK)
+            getContractId(getResult(output))
+          }: Future[ContractId]
+          fut.map(cid => (party, cid))
+        }
+        (alice, aliceUserId) = users(0)
+        (bob, bobUserId) = users(1)
+        _ <- followUser(aliceUserId, alice, bob)
+        _ <- queryUsers(bob)
+        _ <- followUser(bobUserId, bob, alice)
+        _ <- queryUsers(alice)
+      } yield succeed
+  }
 }
